BugFix: SAE Storage file upload failure caused by error handling.
In the code of SAE Storage environment
/usr/local/sae/python/lib/python2.7/site-packages/sae/storage.py in _request:

```
  317                if resp.status < 200 or resp.status >= 300:
* 318                   raise Error(resp.status, resp.reason) ...
  319              resp_headers = {}
```

If the checked file is not exsit, SAE Storage will raise Error not
ObjectNotExistsError while executing self.conn.stat_object(domain, key_name)
in line 423. it's doubting ObjectNotExistsError is instead by Error, or maybe
a bug of SAE Storage, it causes file upload failed.

The patch fixed the issue.